- made streaming read() 'size' arg an 'uint64_t' (was size_t)
   hardware guys had indicated that 64-bit might be required
   somewhere in the distant future...

+++ cpsw_api_user.h
@@ -246,7 +246,7 @@ static const CTimeout TIMEOUT_INDEFINITE;
 
 class IStream {
 public:
-	virtual int64_t read(uint8_t *buf, size_t size, CTimeout timeout = TIMEOUT_INDEFINITE, uint64_t off = 0) = 0;
+	virtual int64_t read(uint8_t *buf, uint64_t size, CTimeout timeout = TIMEOUT_INDEFINITE, uint64_t off = 0) = 0;
 
 
 	static Stream create(Path p);
